Move image.py into `lib/`
